nodes part done, feel free to comment

+++ src/conqueringTheNodes/Displayable.java
@@ -11,8 +11,8 @@ package conqueringTheNodes;
  * @author Admin
  */
 public interface Displayable {
-  public int getX(); // Get x-coordinate of the vertex
-  public int getY(); // Get x-coordinate of the vertex
+  public float getX(); // Get x-coordinate of the vertex
+  public float getY(); // Get x-coordinate of the vertex
   public String getName(); // Get display name of the vertex
 }
