+++ src/conqueringTheNodes/GraphView.java
@@ -15,8 +15,8 @@ public class GraphView extends Pane {
     java.util.List<? extends Displayable> vertices 
       = graph.getVertices();    
     for (int i = 0; i < graph.getSize(); i++) {
-      int x = vertices.get(i).getX();
-      int y = vertices.get(i).getY();
+      float x = vertices.get(i).getX();
+      float y = vertices.get(i).getY();
       String name = vertices.get(i).getName();
       
       getChildren().add(new Circle(x, y, 16)); // Display a vertex
@@ -26,11 +26,11 @@ public class GraphView extends Pane {
     // Draw edges for pairs of vertices
     for (int i = 0; i < graph.getSize(); i++) {
       java.util.List<Integer> neighbors = graph.getNeighbors(i);
-      int x1 = graph.getVertex(i).getX();
-      int y1 = graph.getVertex(i).getY();
+      float x1 = graph.getVertex(i).getX();
+      float y1 = graph.getVertex(i).getY();
       for (int v: neighbors) {
-        int x2 = graph.getVertex(v).getX();
-        int y2 = graph.getVertex(v).getY();
+        float x2 = graph.getVertex(v).getX();
+        float y2 = graph.getVertex(v).getY();
         
         // Draw an edge for (i, v)
         getChildren().add(new Line(x1, y1, x2, y2));
